Move Palette traits to their own directory.
Also adds missing TraitLocation definitions.
