+++ OpenRA.Mods.Cnc/UtilityCommands/RemapShpCommand.cs
@@ -15,8 +15,8 @@ using System.IO;
 using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Cnc.SpriteLoaders;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
-using OpenRA.Traits;
 
 namespace OpenRA.Mods.Cnc.UtilityCommands
 {
