+++ OpenRA.Mods.Common/Traits/Palettes/ColorPickerPalette.cs
@@ -9,7 +9,6 @@
  */
 #endregion
 
-using System;
 using System.Collections.Generic;
 using OpenRA.Graphics;
 using OpenRA.Primitives;
