+++ OpenRA.Mods.Common/Traits/Palettes/IndexedPalette.cs
@@ -15,7 +15,7 @@ using OpenRA.Graphics;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Common
+namespace OpenRA.Mods.Common.Traits
 {
 	[TraitLocation(SystemActors.World | SystemActors.EditorWorld)]
 	[Desc("Define a palette by swapping palette indices.")]
