+++ OpenRA.Mods.Common/Traits/Palettes/PaletteFromFile.cs
