+++ OpenRA.Mods.Common/Traits/Palettes/PaletteFromGimpOrJascFile.cs
