+++ OpenRA.Mods.Common/Traits/Palettes/PaletteFromPaletteWithAlpha.cs
