+++ OpenRA.Mods.Common/Traits/Palettes/PaletteFromPng.cs
