+++ OpenRA.Mods.Common/Traits/Palettes/PaletteFromRGBA.cs
